Fix out-of-bounds entry-chain read in ValueDictionary lookups
The value-type and comparer chain-walk loops in `GetEntryReference` used a
signed `i < entries.Length` guard. An empty bucket encodes as `0`, so the loop
seed `i = GetBucketReference(...) - 1` is `-1`, and `-1 < entries.Length` is
`true` -- the loop then reads `entries[-1]` and walks a garbage chain until it
trips `ThrowForDictionaryConcurrentReadOrWrite`. This reproduces on the very
first `Add`/`TryAdd`/`Remove` against a zero-capacity dictionary (e.g.
`new ValueDictionary<int, int>()` followed by `Add`).

Guard with `(uint)i < (uint)entries.Length` instead, matching both
`UnmanagedValueDictionary` (which already casts) and the unsigned guard in the
`System.Collections.Generic.Dictionary` this is ported from.

+++ sources/Core/Collections/ValueDictionary.cs
@@ -515,7 +515,7 @@ public static class ValueDictionary
                 var hashCode = key.GetHashCode();
                 var collisionCount = 0;
 
-                for (var i = dictionary.GetBucketReference(buckets, hashCode) - 1; i < entries.Length; i = entry.Next)
+                for (var i = dictionary.GetBucketReference(buckets, hashCode) - 1; (uint)i < (uint)entries.Length; i = entry.Next)
                 {
                     entry = ref entries.GetReferenceUnsafe(i);
 
@@ -542,7 +542,7 @@ public static class ValueDictionary
                 var hashCode = comparer.GetHashCode(key);
                 var collisionCount = 0;
 
-                for (var i = dictionary.GetBucketReference(buckets, hashCode) - 1; i < entries.Length; i = entry.Next)
+                for (var i = dictionary.GetBucketReference(buckets, hashCode) - 1; (uint)i < (uint)entries.Length; i = entry.Next)
                 {
                     entry = ref entries.GetReferenceUnsafe(i);
